--- carrot/src/main/java/sopt/carrot/Common/ErrorResponse.java
@@ -1,12 +1,0 @@
-package sopt.carrot.Common;
-
-import sopt.carrot.Exception.Message.ErrorMessage;
-
-public record ErrorResponse(
-        int status,
-        String message
-) {
-    public static ErrorResponse of(int status, String message) {
-        return new ErrorResponse(status, message);
-    }
-}
