--- carrot/src/main/java/sopt/carrot/Common/GlobalExceptionHandler.java
@@ -1,16 +1,0 @@
-package sopt.carrot.Common;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import sopt.carrot.Exception.NotFoundException;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-    @ExceptionHandler(NotFoundException.class)
-    protected ResponseEntity<ErrorResponse> handleNotFoundException(NotFoundException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                .body(ErrorResponse.of(e.getErrorMessage().getStatus(), e.getMessage()));
-    }
-}
