--- carrot/src/main/java/sopt/carrot/Config/JpaAuditingConfig.java
@@ -1,9 +1,0 @@
-package sopt.carrot.Config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-}
