--- carrot/src/main/java/sopt/carrot/Domain/Dto/MemberCreateDto.java
@@ -1,7 +1,0 @@
-package sopt.carrot.Domain.Dto;
-
-public record MemberCreateDto(
-        String name,
-        String phoneNumber
-) {
-}
