--- carrot/src/main/java/sopt/carrot/Domain/Dto/ProductCreateDto.java
@@ -1,12 +1,0 @@
-package sopt.carrot.Domain.Dto;
-
-import sopt.carrot.Domain.TransactionMethod;
-
-public record ProductCreateDto(
-        String title,
-        int price,
-        String description,
-        String location,
-        TransactionMethod transactionMethod
-) {
-}
