--- carrot/src/main/java/sopt/carrot/Domain/TransactionMethod.java
@@ -1,6 +1,0 @@
-package sopt.carrot.Domain;
-
-public enum TransactionMethod {
-    SELL,
-    SHARE
-}
