--- carrot/src/main/java/sopt/carrot/Exception/Message/ErrorMessage.java
@@ -1,14 +1,0 @@
-package sopt.carrot.Exception.Message;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@AllArgsConstructor
-@Getter
-public enum ErrorMessage {
-    MEMBER_NOT_FOUND_BY_ID_EXCEPTION(HttpStatus.NOT_FOUND.value(), "ID에 해당하는 사용자가 존재하지 않습니다.")
-    ;
-    private final int status;
-    private final String messgae;
-}
