--- carrot/src/main/java/sopt/carrot/Exception/NotFoundException.java
@@ -1,9 +1,0 @@
-package sopt.carrot.Exception;
-
-import sopt.carrot.Exception.Message.ErrorMessage;
-
-public class NotFoundException extends BusinessException {
-    public NotFoundException(ErrorMessage errorMessage) {
-        super(errorMessage);
-    }
-}
