--- carrot/src/main/java/sopt/carrot/Repository/ProductRepository.java
@@ -1,7 +1,0 @@
-package sopt.carrot.Repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import sopt.carrot.Domain.Product;
-
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
