+++ carrot/src/main/java/sopt/carrot/common/ErrorResponse.java
@@ -1,0 +1,10 @@
+package sopt.carrot.common;
+
+public record ErrorResponse(
+        int status,
+        String message
+) {
+    public static ErrorResponse of(int status, String message) {
+        return new ErrorResponse(status, message);
+    }
+}
