+++ carrot/src/main/java/sopt/carrot/common/GlobalExceptionHandler.java
@@ -1,0 +1,16 @@
+package sopt.carrot.common;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import sopt.carrot.common.exception.NotFoundException;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+    @ExceptionHandler(NotFoundException.class)
+    protected ResponseEntity<ErrorResponse> handleNotFoundException(NotFoundException e) {
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body(ErrorResponse.of(e.getErrorMessage().getStatus(), e.getMessage()));
+    }
+}
