+++ carrot/src/main/java/sopt/carrot/common/SuccessStatusResponse.java
@@ -1,0 +1,10 @@
+package sopt.carrot.common;
+
+public record SuccessStatusResponse(
+        int status,
+        String message
+) {
+    public static SuccessStatusResponse of(SuccessMessage successMessage) {
+        return new SuccessStatusResponse(successMessage.getStatus(), successMessage.getMessage());
+    }
+}
