+++ carrot/src/main/java/sopt/carrot/common/config/JpaAuditingConfig.java
@@ -1,0 +1,9 @@
+package sopt.carrot.common.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+}
