+++ carrot/src/main/java/sopt/carrot/member/dto/MemberCreateDto.java
@@ -1,0 +1,7 @@
+package sopt.carrot.member.dto;
+
+public record MemberCreateDto(
+        String name,
+        String phoneNumber
+) {
+}
