+++ carrot/src/main/java/sopt/carrot/member/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package sopt.carrot.member.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import sopt.carrot.member.domain.Member;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
