+++ carrot/src/main/java/sopt/carrot/product/domain/TransactionMethod.java
@@ -1,0 +1,6 @@
+package sopt.carrot.product.domain;
+
+public enum TransactionMethod {
+    SELL,
+    SHARE
+}
