+++ carrot/src/main/java/sopt/carrot/product/dto/ProductCreateDto.java
@@ -1,0 +1,12 @@
+package sopt.carrot.product.dto;
+
+import sopt.carrot.product.domain.TransactionMethod;
+
+public record ProductCreateDto(
+        String title,
+        int price,
+        String description,
+        String location,
+        TransactionMethod transactionMethod
+) {
+}
