+++ carrot/src/main/java/sopt/carrot/product/repository/ProductRepository.java
@@ -1,0 +1,7 @@
+package sopt.carrot.product.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import sopt.carrot.product.domain.Product;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+}
